# api/swagger/openapi.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "title": "NATS Horizon API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "version": "1.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/tenancy/connections": {
      "get": {
        "description": "Returns all configured multi-tenancy NATS connections",
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "List tenancy connections",
        "responses": {
          "200": {
            "description": "connections list",
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "post": {
        "description": "Creates a new multi-tenancy NATS connection configuration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Create a tenancy connection",
        "parameters": [
          {
            "description": "Connection configuration",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tenancy/connections/test": {
      "post": {
        "description": "Attempts to connect to a NATS URL and reports connectivity and latency",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Test a connection",
        "parameters": [
          {
            "description": "Connection test request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "connection test result",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tenancy/connections/{id}": {
      "get": {
        "description": "Returns a single configured NATS connection by ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Get a tenancy connection",
        "parameters": [
          {
            "type": "string",
            "description": "Connection ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "description": "Updates an existing multi-tenancy NATS connection configuration",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Update a tenancy connection",
        "parameters": [
          {
            "type": "string",
            "description": "Connection ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Connection configuration",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "description": "Deletes a multi-tenancy NATS connection configuration (the default connection cannot be deleted)",
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Delete a tenancy connection",
        "parameters": [
          {
            "type": "string",
            "description": "Connection ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tenancy/connections/{id}/default": {
      "get": {
        "description": "Marks a connection as the default NATS connection",
        "produces": [
          "application/json"
        ],
        "tags": [
          "tenancy"
        ],
        "summary": "Set default connection",
        "parameters": [
          {
            "type": "string",
            "description": "Connection ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ConnectionConfig"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ConnectionConfig": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "id": {
          "type": "string"
        },
        "is_default": {
          "type": "boolean"
        },
        "name": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
  }
}